+++ apps/3002-checkout/src/components/menu.tsx
