+++ packages/node/src/utils/flush-chunks.js
@@ -29,6 +29,7 @@ export const flushChunks = async () => {
         }
       })
     }
+
     return Array.from(chunks)
   }))
